+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT RELEASE
 * Clear `TheAlgorithm.cacheUpdater` in `reset()`
 * Rename `Config` to `config`; convert `config.api.staleDataTrendingMinutes` to method
+* Rename config.api.backgroundLoadIntervalSeconds to backgroundLoadIntervalMinutes
 
 ### v0.47.8
 * Make `Config` a `class` instead of just an object
@@ -497,7 +498,7 @@
 * Privatize a bunch of `TheAlgorithm` properties that should be private
 
 ### v0.24.1
-* Lower `Config.maxRecordsForFeatureScoring` to 1600, `numTrendingTagsToots` to 200
+* Lower `Config.maxRecordsForFeatureScoring` to 1,600, `numTrendingTagsToots` to 200
 * `HashtagParticipationScorer` uses the square root of the number of participations to tamp down runaway scores.
 * Use a Semaphore to limit concurrency of hashtag search API requests; add `Config.maxConcurrentTootRequests`
 * Exclude followed tags from the pulling of hashtag participation toots (followed tags are already in timeline)